Add support for organisations (#141)
* Add organisations to the home page
Add a home page component. Add an action to retrieve organisations from Panoptes. Add an organisation list to the home page.

* Rename ProjectContentsContainer
This is used as a generic container for any resource, so rename it.

* Add an organisation editor
Add an organisation editor, copied from the project editor.

* Pass primary_language down to resources
Pass organisation or project primary language down, so that it can be used when fetching translations.

* Wire up the organisation reducers

* Set resource type and ID for organisations

* Tidy up Resource components
Always pass resource type and id as params.

* Add admin mode for organisations